+++ be/promotion/src/test/java/woowa/promotion/util/fixture/FixtureFactory.java
@@ -10,8 +10,8 @@ import woowa.promotion.admin.coupon.domain.Coupon;
 import woowa.promotion.admin.coupon.domain.CouponType;
 import woowa.promotion.admin.coupon_group.domain.CouponGroup;
 import woowa.promotion.admin.coupon_group.domain.Type;
-import woowa.promotion.admin.coupon_group.presentation.dto.CouponGroupCreateRequest;
-import woowa.promotion.admin.coupon_group.presentation.dto.CouponGroupCreateRequest.CouponDto;
+import woowa.promotion.admin.coupon_group.presentation.dto.request.CouponGroupCreateRequest;
+import woowa.promotion.admin.coupon_group.presentation.dto.request.CouponGroupCreateRequest.CouponDto;
 import woowa.promotion.admin.promotion.application.dto.request.PromotionRegisterRequest;
 import woowa.promotion.admin.promotion.application.dto.request.PromotionRegisterRequest.PromotionOptionRequest;
 import woowa.promotion.admin.promotion.domain.ProgressStatus;
